Finish pitcher strikeout average calculator
After user inputs "1" in getPlayerPosition() and getCategory()
functions, they will be prompted with how many strikeouts the pitcher
had and over the course of how many games.

If the user does not enter a valid number, then they will be reprompted
the assignment. (They will have 5 tries before the script closes)

After inputting the strikeouts and games, the average will be calculated
in the calcStrikeouts() function.

Finally, the user will be asked if they would like to do more
calculations with the same player or a new one.